add basic CLI commands, split the schema up into different files
