+++ asapdiscovery-simulation/asapdiscovery/simulation/cli/cli.py
@@ -1,0 +1,21 @@
+import click
+
+from asapdiscovery.simulation.cli.create import create
+from asapdiscovery.simulation.cli.plan import plan
+from asapdiscovery.simulation.cli.submit import submit
+from asapdiscovery.simulation.cli.status import status
+
+
+@click.group()
+def cli():
+    pass
+
+
+cli.add_command(create)
+cli.add_command(plan)
+cli.add_command(submit)
+# cli.add_command(results)
+cli.add_command(status)
+
+if __name__ == "__main__":
+    cli()
