+++ asapdiscovery-simulation/asapdiscovery/simulation/cli/submit.py
@@ -1,0 +1,43 @@
+import click
+
+
+@click.command()
+@click.option("-n", "--network", type=click.Path(resolve_path=True, readable=True, file_okay=True, dir_okay=False), help="The name of the JSON file containing a planned FEC network.", default="planned_network.json", show_default=True)
+@click.option("-o", "--organization", type=click.STRING, default="asap", help="The name of the organization in alchemiscale the network should be submitted to.", show_default=True)
+@click.option("-c", "--campaign", type=click.STRING, help="The name of the campaign in alchemiscale the network should be submitted to.", required=True)
+@click.option("-p", "--project", type=click.STRING, help="The name of the project in alchemiscale the network should be submitted to.", required=True)
+def submit(network: str, organization: str, campaign: str, project: str):
+    """
+    Submit a local FreeEnergyCalculationNetwork to alchemiscale using the provided scope details. The network object
+    will have these details saved into it.
+
+    Args:
+        network: The name of the JSON file containing the FreeEnergyCalculation to be submitted.
+        organization: The name of the organization this network should be submitted under always asap.
+        campaign: The name of the campaign this network should be submitted under.
+        project: The name of the project this network should be submitted under.
+    """
+    from asapdiscovery.simulation.utils import AlchemiscaleHelper
+    from asapdiscovery.simulation.schema.fec import FreeEnergyCalculationNetwork
+    from alchemiscale import Scope
+
+    # launch the helper which will try to login
+    click.echo("Connecting to Alchemiscale...")
+    client = AlchemiscaleHelper()
+    # create the scope
+    network_scope = Scope(org=organization, campaign=campaign, project=project)
+    # load the network
+    planned_network = FreeEnergyCalculationNetwork.from_file(network)
+    # create network on alchemiscale
+    click.echo(f"Creating network on Alchemiscale instance: {client._client.api_url} with scope {network_scope}")
+    submitted_network = client.create_network(planned_network=planned_network, scope=network_scope)
+    # write the network with its key to file before we try and add compute incase we hit an issue
+    click.echo("Network made saving network key to network file")
+    submitted_network.to_file("planned_network.json")
+    # now action the tasks
+    click.echo("Creating and actioning FEC tasks on Alchemiscale...")
+    task_ids = client.action_network(planned_network=submitted_network)
+    # check that all tasks where created
+    missing_tasks = sum([1 for task in task_ids if task is None])
+    total_tasks = len(task_ids)
+    click.echo(f"{total_tasks - missing_tasks}/{total_tasks} created. Status can be checked using `asap-fec status`")
