+++ asapdiscovery-simulation/pyproject.toml
@@ -29,6 +29,9 @@ dependencies = [
 "Homepage" = "https://github.com/choderalab/asapdiscovery"
 "Bug Tracker" = "https://github.com/choderalab/asapdiscovery/issues"
 
+[project.scripts]
+asap-fec = "asapdiscovery.simulation.cli.cli:cli"
+
 [tool.setuptools.packages.find]
 where = ["."]
